+++ calculationbases/cost/acquisitioncosts/AcquisitionCostRates.csv
