+++ calculationbases/cost/acquisitioncosts/acquisition_rate.py
@@ -1,12 +1,12 @@
 from os import path
-from CPL_Prep import FileReader
+from helper.cpl_prep import FileReader
 from CalculationBases.Costs.CostMapping import CostMapping
 
 
-class UnitCost:
+class AcquisitionRate:
     """
-    Class used to provide the unit cost rate. Its method relies on a csv.
-    UnitCostRates.csv links the unit cost rate to the unit cost group coming from the class CostMapping.py.
+    Class used to provide the acquisition cost rate. Its method relies on a csv.
+    AcquisitionRates.csv links the unit cost rate to the acquisition cost group coming from the class cost_mapping.py.
     """
 
     def __init__(self, contract_nr):  # ToDo: Logic for CostGroups needs to be included, or be part of the IL
@@ -14,17 +14,17 @@ class UnitCost:
         :param int contract_nr: The mapping is done via the contract number, which provides the tariff name
         """
         relative_path = "/"
-        csv_filename = "UnitCostRates.csv"
+        csv_filename = "AcquisitionCostRates.csv"
         csv_path = path.dirname(__file__) + relative_path + csv_filename
         csv_reader = FileReader(csv_path)
-        self.mapping_dictionary = csv_reader.create_mapping_by_key("UnitCostGroups")
-        self.cost_group = CostMapping(contract_nr=contract_nr).unit_cost_group()
+        self.mapping_dictionary = csv_reader.create_mapping_by_key("AcquisitionCostGroups")
+        self.cost_group = CostMapping(contract_nr=contract_nr).acquisition_cost_group()
 
-    def get_unit_cost_by_name(self, cost_type):
+    def get_acquisition_cost_by_name(self, cost_type):
         """
-        Getting the unit cost rate for a cost type, based on cost group
+        Getting the acquisitions cost rate for a cost type, based on cost group
         :param cost_type: Needs to match the cost_type name in the csv
-        :return: Unit cost rate
+        :return: Acquisition cost rate
         """
         value = self.mapping_dictionary[self.cost_group][cost_type]
         return float(value)
